upgrade applications to v6.0

+++ android/test-app-local-dependency/src/main/java/com/arthenica/ffmpegkit/test/PipeTabFragment.java
@@ -250,7 +250,7 @@ public class PipeTabFragment extends Fragment {
             return;
         }
 
-        int timeInMilliseconds = this.statistics.getTime();
+        double timeInMilliseconds = this.statistics.getTime();
         int totalVideoDuration = 9000;
 
         String completePercentage = new BigDecimal(timeInMilliseconds).multiply(new BigDecimal(100)).divide(new BigDecimal(totalVideoDuration), 0, BigDecimal.ROUND_HALF_UP).toString();
