+++ macos/test-app-local-dependency/FFmpegKitMACOS/SubtitleViewController.m
@@ -218,7 +218,7 @@ typedef enum {
         return;
     }
 
-    int timeInMilliseconds = [statistics getTime];
+    double timeInMilliseconds = [statistics getTime];
     int totalVideoDuration = 9000;
 
     int percentage = timeInMilliseconds*100/totalVideoDuration;
